adding class to store coordinates
