--- test_input_ddm.csv
@@ -1,13 +1,0 @@
-ID,species,latitude,longitude
-1,TT,32° 30.808 N,085° 36.678W
-2,TT,32° 30.851 N,085° 36.799 W
-3,TT,32° 30.771 N,085° 36.720 W
-4,TT,32° 30.848 N,085° 36.683 W
-5,TT,32° 30.837 N,085° 36.765 W
-6,EA,32° 30.788 N,085° 36.775 W
-7,EA,32° 30.757 N,085° 36.815 W
-8,EA,32° 30.836 N,085° 36.773 W
-9,EA,32° 30.747 N,085° 36.795 W
-10,EA,32° 30.729 N,085° 36.627 W
-11,EA,32° 30.786 N,085° 36.776 W
-12,EA,32° 30.772 N,085° 36.710W
