--- test_input_dms.csv
@@ -1,13 +1,0 @@
-ID,species,latitude,longitude
-1,TT,N 32 32 263.32,W -85 30 253
-2,TT,N 32 41 219.32,W -85 19 237
-3,TT,N 32 43 207.34,W -85 18 101
-4,TT,N 32 46 205.56,W -85 15 542
-5,TT,N 32 47 919.65,W -85 14 002
-6,EA,N 32 48 308.52,W -85 13 663
-7,EA,N 32 43 168,W -85 18 152
-8,EA,N 32 46 797,W -85 34 556
-9,EA,N 32 32 194,W -85 29 902
-10,EA,N 32 34 314,W -85 31 828
-11,EA,  32 32 247,W -85 30 369
-12,EA,  32 32 144,85 29 753 W
